Some compatibility work for the simulated data
The simulate_slice_data function now outputs in a more appropriate format for the BayesGLM_slice function. Some namespace work was done to eliminate a note that we were getting from R CMD CHECK about rewriting functions from the stats and dplyr packages. The plot_image_slice (and it's containing script) were renamed plot_slice. A small change was made to BayesGLM_slice to account for the new simulated data format (which was made to match the data input format on BayesGLM_slice)